feat: 🎸 loadApplication完善
