+++ packages/core/src/common/index.ts
@@ -5,7 +5,7 @@ export { default as zip } from './zip';
 export { default as unzip } from 'decompress';
 export { default as i18n } from './i18n';
 export { default as help } from './help';
-export { default as loadComponent, load } from './load';
+export { loadComponent, load, loadApplication } from './load';
 export { default as commandParse } from './commandParse';
 export { getCredential, setCredential } from './credential';
 export { getState, setState } from './state';
