+++ packages/core/test/fixtures/load.ts
@@ -1,11 +1,11 @@
 import { loadComponent } from '../../src/common';
 
 async function test() {
-  // const [FC] = await loadComponent('alibaba/fc', 'https://tool.serverlessfans.com/api');
-
+  const [FC] = await loadComponent('alibaba/fc', 'https://tool.serverlessfans.com/api');
   // const [FC] = await loadComponent('Serverless-Devs/Serverless-Devs@1.1.13');
-  const [FC] = await loadComponent('Serverless-Devs/Serverless-Devs');
+  // const [FC] = await loadComponent('Serverless-Devs/Serverless-Devs');
   console.log(FC);
+  // await loadApplication('alibaba/fc', 'https://tool.serverlessfans.com/api');
 }
 
 test();
